Automated pattern analysis
The code extracts the tables from a database and according to some configuration automatically, selects tables, filters, normalizes it (joins¡ing tables), aggregates in different ways, imputes missing data and runs advanced descriptive methods.
The methods used are the biplot, a method similar to PCA that allows to create an optimal representation of both the rows and columns to see the main patterns in the data.